# src/main/java/ar/edu/utn/frc/tup/lc/iv/models/Distrito.java
package ar.edu.utn.frc.tup.lc.iv.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Distrito {
    private int distritoId;
    private String distritoNombre;
}